+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Upcoming
+* [FEATURE] display a badge count next to the application icon
+
 ## v1.0.0-alpha.5 - 2017-06-12
 
 * [MAINTENANCE] add new package mode that does not bundle any DeskPro sdk, drop peer dependencies